// src/store/notification/notification.enum.ts
enum NotificationActionTypeEnum {
  GET_NOTIFICATION_SUCCESS = "GET_NOTIFICATION_SUCCESS",
  GET_NOTIFICATION_FAILED = "GET_NOTIFICATION_FAILED",
  GET_NOTIFICATION_PENDING = "GET_NOTIFICATION_PENDING",

  GET_NOTIFIED_USERS_SUCCESS = "GET_NOTIFIED_USER_SUCCESS",
  GET_NOTIFIED_USERS_FAILED = "GET_NOTIFIED_USER_FAILED",
  GET_NOTIFIED_USERS_PENDING = "GET_NOTIFIED_USER_PENDING",

  SEND_NOTIFICATION_SUCCESS = "SEND_NOTIFICATION_SUCCESS",
  SEND_NOTIFICATION_FAILED = "SEND_NOTIFICATION_FAILED",
  SEND_NOTIFICATION_PENDING = "SEND_NOTIFICATION_PENDING",

  DELETE_NOTIFICATION_SUCCESS = "DELETE_NOTIFICATION_SUCCESS",
  DELETE_NOTIFICATION_FAILED = "DELETE_NOTIFICATION_FAILED",
  DELETE_NOTIFICATION_PENDING = "DELETE_NOTIFICATION_PENDING",

  UPDATE_NOTIFICATION_SUCCESS = "UPDATE_NOTIFICATION_SUCCESS",
  UPDATE_NOTIFICATION_FAILED = "UPDATE_NOTIFICATION_FAILED",
  UPDATE_NOTIFICATION_PENDING = "UPDATE_NOTIFICATION_PENDING",
}

export default NotificationActionTypeEnum;